requirements.txt
mcp
python-docx
requests
